+++ changelog.md
@@ -1,1 +1,1 @@
-- Added bind address fallback
+- Fixed horizontal group layout