+++ changelog.md
@@ -1,1 +1,1 @@
-- Fixed horizontal group layout
+- Updated to 1.17.1-pre3